+++ app/src/main/java/com/sopt/gongbaek/presentation/ui/auth/screen/MajorSearchScreen.kt
@@ -1,4 +1,4 @@
-package com.sopt.gongbaek.presentation.ui.onboarding.screen
+package com.sopt.gongbaek.presentation.ui.auth.screen
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
@@ -12,12 +12,12 @@ import androidx.compose.ui.tooling.preview.Preview
 import com.sopt.gongbaek.ui.theme.GONGBAEKTheme
 
 @Composable
-fun SelfIntroductionRoute() {
-    SelfIntroductionScreen()
+fun MajorSearchRoute() {
+    MajorSearchScreen()
 }
 
 @Composable
-fun SelfIntroductionScreen() {
+fun MajorSearchScreen() {
     Box(
         modifier = Modifier
             .fillMaxSize()
@@ -25,15 +25,15 @@ fun SelfIntroductionScreen() {
         contentAlignment = Alignment.Center
     ) {
         Text(
-            text = "자기소개 화면"
+            text = "학과 검색 화면"
         )
     }
 }
 
 @Preview(showBackground = true)
 @Composable
-fun ShowSelfIntroductionScreen() {
+fun ShowMajorSearchScreen() {
     GONGBAEKTheme {
-        SelfIntroductionScreen()
+        MajorSearchScreen()
     }
 }
