+++ app/src/main/java/com/sopt/gongbaek/presentation/ui/auth/screen/UnivMajorScreen.kt
@@ -1,4 +1,4 @@
-package com.sopt.gongbaek.presentation.ui.onboarding.screen
+package com.sopt.gongbaek.presentation.ui.auth.screen
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
@@ -12,12 +12,12 @@ import androidx.compose.ui.tooling.preview.Preview
 import com.sopt.gongbaek.ui.theme.GONGBAEKTheme
 
 @Composable
-fun NicknameRoute() {
-    NicknameScreen()
+fun UnivMajorRoute() {
+    UnivMajorScreen()
 }
 
 @Composable
-fun NicknameScreen() {
+fun UnivMajorScreen() {
     Box(
         modifier = Modifier
             .fillMaxSize()
@@ -25,15 +25,15 @@ fun NicknameScreen() {
         contentAlignment = Alignment.Center
     ) {
         Text(
-            text = "닉네임 입력 화면"
+            text = "학교/학과 입력 화면"
         )
     }
 }
 
 @Preview(showBackground = true)
 @Composable
-fun ShowNicknameScreen() {
+fun ShowUnivMajorScreen() {
     GONGBAEKTheme {
-        NicknameScreen()
+        UnivMajorScreen()
     }
 }
