--- app/src/main/java/com/sopt/gongbaek/presentation/ui/onboarding/navigation/OnboardingNavigator.kt
@@ -1,12 +1,0 @@
-package com.sopt.gongbaek.presentation.ui.onboarding.navigation
-
-import androidx.navigation.NavController
-import com.sopt.gongbaek.presentation.model.NavigationRoute
-
-fun NavController.navigateOnboardingRoute() {
-    navigate(route = NavigationRoute.OnboardingNavGraphRoute.ONBOARDING_NAV_GRAPH)
-}
-
-fun NavController.navigateCompleteOnboarding() {
-    navigate(route = NavigationRoute.OnboardingNavGraphRoute.ONBOARDING)
-}
